requirements.txt
fiona==1.9.5
GDAL==3.8.4
geopandas==1.0.1
numpy==2.2.6
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.2.1
PySide6==6.8.0.2
PySide6==6.9.0
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
rasterio==1.3.9
setuptools==75.1.0
Shapely==2.1.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
ultralytics==8.3.27
